requirements.txt
discord.py[voice]
requests
youtube_dl
gtts
google-api-python-client
